Add a new frontend router
