+++ packages/backend/src/config/Config.ts
@@ -16,4 +16,5 @@ export interface Config {
   freshStart: boolean
   forceHttps: boolean
   starkex: StarkexConfig
+  useOldFrontend: boolean
 }
