+++ packages/backend/src/config/environments/config.local.ts
@@ -20,5 +20,6 @@ export function getLocalConfig(): Config {
     freshStart: getEnv.boolean('FRESH_START', false),
     forceHttps: false,
     starkex: getStarkexConfig(getEnv('STARKEX_INSTANCE')),
+    useOldFrontend: false,
   }
 }
