+++ packages/backend/src/config/environments/config.production.ts
@@ -20,5 +20,6 @@ export function getProductionConfig(): Config {
     freshStart: false,
     forceHttps: true,
     starkex: getStarkexConfig(getEnv('STARKEX_INSTANCE')),
+    useOldFrontend: true,
   }
 }
